extract into files
